Register endpoint ok

+++ src/main/java/ch/benoitschopfer/repository/ContactRepository.java
@@ -4,7 +4,5 @@ import ch.benoitschopfer.model.Contact;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-import java.math.BigDecimal;
-
-public interface ContactRepository extends JpaRepository<Contact, BigDecimal>, CrudRepository<Contact, BigDecimal> {
+public interface ContactRepository extends JpaRepository<Contact, Long>, CrudRepository<Contact, Long> {
 }
