+++ src/main/java/ch/benoitschopfer/repository/UserRepository.java
@@ -4,7 +4,5 @@ import ch.benoitschopfer.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-import java.math.BigDecimal;
-
-public interface UserRepository extends JpaRepository<User, BigDecimal>, CrudRepository<User, BigDecimal> {
+public interface UserRepository extends JpaRepository<User, Long>, CrudRepository<User, Long> {
 }
